# Optimize ROC AUC Computation