--- a/integration_tests/benchmarks/classification/benchmark_script.py
+++ b/integration_tests/benchmarks/classification/benchmark_script.py
@@ -156,7 +156,7 @@ def run_detailed_pr_curve_evaluation(dset: Dataset, model: Model):
 
 
 def run_benchmarking_analysis(
-    limits_to_test: list[int] = [1000, 1000],
+    limits_to_test: list[int] = [5000, 5000],
     results_file: str = "results.json",
     data_file: str = "data.json",
 ):
@@ -193,6 +193,22 @@ def run_benchmarking_analysis(
                 f"Evaluation timed out when processing {limit} datums."
             )
 
+        try:
+            eval_pr = run_pr_curve_evaluation(dset=dset, model=model)
+        except TimeoutError:
+            raise TimeoutError(
+                f"Evaluation timed out when processing {limit} datums."
+            )
+
+        try:
+            eval_pr_detail = run_detailed_pr_curve_evaluation(
+                dset=dset, model=model
+            )
+        except TimeoutError:
+            raise TimeoutError(
+                f"Evaluation timed out when processing {limit} datums."
+            )
+
         start = time.time()
         client.delete_dataset(dset.name, timeout=30)
         client.delete_model(model.name, timeout=30)
@@ -204,6 +220,8 @@ def run_benchmarking_analysis(
             "number_of_annotations": eval_.meta["annotations"],
             "ingest_runtime": f"{(ingest_time):.1f} seconds",
             "eval_runtime": f"{(eval_.meta['duration']):.1f} seconds",
+            "eval_pr_runtime": f"{(eval_pr.meta['duration']):.1f} seconds",
+            "eval_pr_detail_runtime": f"{(eval_pr_detail.meta['duration']):.1f} seconds",
             "del_runtime": f"{(deletion_time):.1f} seconds",
         }
         write_results_to_file(write_path=write_path, result_dict=results)